requirements.txt
requests==2.32.3
pandas==2.2.3
sqlalchemy==2.0.36
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1